feat : 시간 사용하는 예약 존재하는 경우 서비스에서 예외 처리

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -5,6 +5,7 @@ import java.time.format.DateTimeParseException;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import roomescape.domain.ReservationRepository;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.ReservationTimeRepository;
 import roomescape.dto.app.ReservationTimeAppRequest;
@@ -13,10 +14,13 @@ import roomescape.dto.app.ReservationTimeAppRequest;
 public class ReservationTimeService {
 
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
     @Autowired
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository,
+        ReservationRepository reservationRepository) {
         this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public ReservationTime save(ReservationTimeAppRequest request) {
@@ -30,6 +34,10 @@ public class ReservationTimeService {
     }
 
     public void delete(Long id) {
+        long count = reservationRepository.countByTimeId(id);
+        if (count > 0) {
+            throw new IllegalArgumentException("해당 시간을 사용하는 예약이 존재합니다.");
+        }
         reservationTimeRepository.deleteById(id);
     }
